CLOUD-338 removed unused domain object
